Fetch and clone git sources automatically.
This commit adds a clever way for dealing with git based packages. Since
most public git servers do not support git-archive, we must use git-clone
instead. We simply clone the git projects into a local directory given as
SLIM_GIT in the environment. If this directory does not exist, we assume
the variable points to a local git server allowing git-archive.

Signed-off-by: Richard Cochran <richard.cochran@omicron.at>