[PKG] Reject the defaults for new busybox config options.
When building a busybox that has new options not in our .config, the
build will block on 'silentoldconfig' waiting for user input. Since
the defaults are almost always 'yes', the defaults may lead to
unwanted bloat. Therefore, we just say 'no', even though this runs the
risk of refusing critical new options.

Reported-by: Jens Steinhauser <jens.steinhauser@gmail.com>
Signed-off-by: Richard Cochran <richardcochran@gmail.com>