[pkg] busybox: fix the configuration file handling.
Busybox does not have an 'oldnoconfig' target or any reasonable
alternative for it. When updating an older config file, the best that we
can do automatically is to accept the new defaults.

Saying "n" to new options, as we tried before, will not work because
some options are not simple yes and no questions. For example, there can
be numbered choices.

Also, we should do either 'menuconfig' or 'oldconfig' but not both. In that
way, if the user does make K=1, then he will see the new options clearly
marked.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>