Fix multiarch glibc installation.
With the most recent linaro arm gnueabihf tool chain, the scp
service does not work.  When connecting to the target, the
following error appears.

   scp: error while loading shared libraries:
   libcrypto.so.1.0.0: cannot open shared object file:
   No such file or directory

That library is in fact installed, in /usr/lib, and so there is a
problem with the LDO search paths in the multiarch loader.

This issue can be resolved in a number of ways:

1. Make the libraries install into /usr/lib/<arch>.  This would
   be a PITA since it involves changing every last makefile in every
   package that delivers a LDO.

2. Installing ldconfig and running it at least once on the target.  We
   definitely don't want this, because that program is statically
   linked and huge in size.

3. Setting a link in /lib/<arch>.  When running ldconfig on the
   target, it links ld-linux-armhf.so.3 -> ld-2.19-2014.08.so in
   /lib/arm-linux-gnueabihf.  Why this is missing from linaro and why
   this would solve the issue is beyond me.  Interestingly, debian
   does have that link by default.  It appears that the link should be
   set both in /lib and in /lib/<arch>.

This patch takes the easy way out and simulates #1 by setting a
symbolic link, making the contents of /usr/lib appear in
/usr/lib/<arch> as well.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>