Touch the source files when unpacking an archive.
The SLIM system will rebuild a package whenever its makefile is changed.
However, the time stamps of the newly extracted source files could very
well be before the old build products.  As a result, stale binaries might
be re-installed into the target file system.

This patch fixes the issue by forcing the time stamps on the extracted
files to the current system time.

Signed-off-by: Richard Cochran <richardcochran@gmail.com>